+++ src/main/java/com/example/timesaver/model/dto/team/TeamNrMembers.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class TeamNrMembers {
     private Team team;
-    private Long nrOfMembers;
+    private Integer nrOfMembers;
 }
