+++ src/main/java/com/example/timesaver/model/dto/teamflow/TeamApplicationDTO.java
@@ -4,9 +4,9 @@ import com.example.timesaver.model.TeamApplication.Status;
 import java.time.ZonedDateTime;
 
 public record TeamApplicationDTO(
-        Long id,
-        Long teamId,
-        Long applicantId,
+        Integer id,
+        Integer teamId,
+        Integer applicantId,
         String firstName,
         String lastName,
         Status status,
