+++ src/main/java/com/example/timesaver/repository/TeamBackgroundRequirementRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TeamBackgroundRequirementRepository extends JpaRepository<TeamBackgroundRequirement, Long> {
+public interface TeamBackgroundRequirementRepository extends JpaRepository<TeamBackgroundRequirement, Integer> {
     List<TeamBackgroundRequirement> findByTeam(Team team);
 }
